requirements.txt
azure-synapse
azure-identity
azure-synapse-artifacts
